# coletor_go/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Motor de Ingestão e Gestão Patrimonial.",
        "title": "QuantAdvisor API",
        "contact": {},
        "version": "1.0"
    },
    "host": "100.95.28.45:8080",
    "basePath": "/api",
    "paths": {
...
        "/agente/causalidade": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Encaminha prompts de raciocínio lógico financeiro para a IA Generativa (LLM). O modelo cruza os logs internos do motor estatístico com os dados macroeconômicos do mercado para explicar as decisões da rede neural de forma didática.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "IA"
                ],
                "summary": "Agente Causalidade",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}